Refactoring team formation
