+++ WhenPushComesToShove/Assets/1-Scenes/TestScenes/DestinyScene.unity
@@ -2026,6 +2026,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1755526682}
   - component: {fileID: 1755526681}
+  - component: {fileID: 1755526683}
   - component: {fileID: 1755526680}
   m_Layer: 0
   m_Name: Player Config Manager
@@ -2133,9 +2134,6 @@ MonoBehaviour:
   - {r: 0, g: 0.84313726, b: 0.99607843, a: 1}
   - {r: 0, g: 0.8627451, b: 0.09019608, a: 1}
   - {r: 0.99607843, g: 0.8117647, b: 0, a: 1}
-  teamSprites:
-  - {fileID: 21300000, guid: 3761eb0b08fa2ba4ca8e488badd09078, type: 3}
-  - {fileID: 21300000, guid: 4cee1458bbddae9418c3bf59117a05ed, type: 3}
 --- !u!4 &1755526682
 Transform:
   m_ObjectHideFlags: 0
@@ -2151,6 +2149,21 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1755526683
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1755526679}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 97719b6d59edf9844ad5b4b8db6254fa, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  teamSprites:
+  - {fileID: 21300000, guid: 3761eb0b08fa2ba4ca8e488badd09078, type: 3}
+  - {fileID: 21300000, guid: 4cee1458bbddae9418c3bf59117a05ed, type: 3}
 --- !u!1 &1787772874
 GameObject:
   m_ObjectHideFlags: 0
